% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_dayperiod}
\alias{add_dayperiod}
\title{Add column with period of the day to a tibble with time column "t"}
\usage{
add_dayperiod(df)
}
\arguments{
\item{df}{A tibble with datetime or time column 't'}
}
\description{
Add column with period of the day to a tibble with time column "t"
}
