% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{all_paths}
\alias{all_paths}
\title{Search for all simple paths from source to sink. This is a wrapper around
igraph::all_simple_paths that tidys up the results into a single tibble.}
\usage{
all_paths(G)
}
\arguments{
\item{G}{a flow graph}
}
\value{
a tibble
}
\description{
Search for all simple paths from source to sink. This is a wrapper around
igraph::all_simple_paths that tidys up the results into a single tibble.
}
