% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_tuples}
\alias{as_tuples}
\title{Transform a vector of n elements (1,2,3) into a vector of consecutive
character tuples of length n-1 ("1-2","2-3").}
\usage{
as_tuples(v, unite_sep = "-")
}
\arguments{
\item{v}{vector}

\item{unite_sep}{character separator to unite consecutive elements into tuples}
}
\description{
Transform a vector of n elements (1,2,3) into a vector of consecutive
character tuples of length n-1 ("1-2","2-3").
}
