% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_cols}
\alias{assert_cols}
\title{Assert that cols are present in input tibble}
\usage{
assert_cols(tib, required_cols)
}
\arguments{
\item{tib}{a tibble}

\item{required_cols}{character vector of columns expected to exist in tib}
}
\value{
logical if assert is true
}
\description{
Assert that cols are present in input tibble
}
