% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_tidygraph}
\alias{assert_tidygraph}
\title{Assert that input is a tidygraph}
\usage{
assert_tidygraph(x)
}
\arguments{
\item{x}{input}
}
\value{
logical
}
\description{
Assert that input is a tidygraph
}
