% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{count_sequences}
\alias{count_sequences}
\title{Get and count trip sequences}
\usage{
count_sequences(df, sep = ",")
}
\arguments{
\item{df}{trips dataset with three columns: (k,i,x) i-th trip of vehicle k is sequence of x's}

\item{sep}{sequence separator character}
}
\value{
tibble with columns (s = sequence, l = length of sequence, n = count)
}
\description{
Get and count trip sequences
}
