% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{crop_spatial}
\alias{crop_spatial}
\title{Plot a map of}
\usage{
crop_spatial(
  flows_od,
  spatial,
  arterial_highway = "residential",
  bbox_margin = c(-250, -250, 250, 250)
)
}
\arguments{
\item{flows_od}{Trimmed flows$od tibble.}

\item{spatial}{List of spatial features.}

\item{arterial_highway}{Only keep arterial edges with this value in the
highway attribute.}

\item{bbox_margin}{Spatial margin added to flows bounding box.}
}
\description{
Plot a map of
}
