% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{cross_super_sequences}
\alias{cross_super_sequences}
\title{Compute the cross product of one set of trip sequences and annotate them}
\usage{
cross_super_sequences(super_sequences)
}
\arguments{
\item{super_sequences}{tibble of ordinary sequences}
}
\value{
expanded tibble of super_sequences
}
\description{
Compute the cross product of one set of trip sequences and annotate them
}
