% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{cut_flows}
\alias{cut_flows}
\title{Get the subset of flows whose OD pairs account for the top
p proportion of total or period-wise observed traffic flow.
Returns all od pairs up to the first od pair that exceeds p.}
\usage{
cut_flows(
  flows_od,
  flows_l,
  time_resolution,
  locations = NULL,
  pairs = NULL,
  filter_dayperiod = "morning",
  fill_gaps = TRUE,
  filter_weekdays = c(1:5),
  min_t = NULL,
  max_t = NULL
)
}
\arguments{
\item{flows_od}{flows at ods [tibble][tibble::tibble-package]}

\item{flows_l}{flows at locations [tibble][tibble::tibble-package]}

\item{time_resolution}{temporal resolution of flow tibbles}

\item{locations}{filter observations occurring at these locations (vector)}

\item{pairs}{filter observations involving these
od pairs [tibble][tibble::tibble-package]}

\item{filter_dayperiod}{filter observations occurring
at this period of the day}

\item{fill_gaps}{whether to make missing time steps explicit}

\item{filter_weekdays}{filter observations occurring on these weekdays}

\item{min_t}{first expected time step, defaults to minimum time in data}

\item{max_t}{last expected time step, defaults to maximum time in data}
}
\value{
list with updated tibbles of flows_od and flows_l
}
\description{
Get the subset of flows whose OD pairs account for the top
p proportion of total or period-wise observed traffic flow.
Returns all od pairs up to the first od pair that exceeds p.
}
