% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dayperiod}
\alias{dayperiod}
\title{Categorise time into its corresponding period of the day}
\usage{
dayperiod(t)
}
\arguments{
\item{t}{time vector}
}
\value{

}
\description{
Categorise time into its corresponding period of the day
}
