% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{expand_sequences}
\alias{expand_sequences}
\title{Expand any trip sequence of length 3 or greater, with their subsequences
of length 2 or greater.}
\usage{
expand_sequences(seqs, sep = ",")
}
\arguments{
\item{seqs}{a tibble with named column "s": sequence}

\item{sep}{character separating each element in the sequence}
}
\value{
expanded tibble, with additional column "l" specifying the length of
the sequence
}
\description{
For example, given sequence "A,B,C,D", we generate all of its subsequences
of length 2 and 3, resulting in the following sequence set:
"A,B", "B,C", "C,D", "A,B,C", "B,C,D", "A,B,C,D"
}
