% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{fill_time_gaps}
\alias{fill_time_gaps}
\title{Make missing time steps explicit (zero vehicle count).}
\usage{
fill_time_gaps(flows_od, flows_l, expected_time)
}
\arguments{
\item{flows_od}{flows at ods [tibble][tibble::tibble-package]}

\item{flows_l}{flows at locations [tibble][tibble::tibble-package]}

\item{expected_time}{[tibble][tibble::tibble-package] with expected steps}
}
\value{
list with updated tibbles of flows_od and flows_l
}
\description{
Make missing time steps explicit (zero vehicle count).
}
