% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_element_greater}
\alias{first_element_greater}
\title{Get first element greater than number in vector.}
\usage{
first_element_greater(vec, number)
}
\arguments{
\item{vec}{numeric}

\item{number}{numeric of length 1}
}
\description{
Get first element greater than number in vector.
}
