% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{flow_edges}
\alias{flow_edges}
\title{Get edges tibble from flow network}
\usage{
flow_edges(network, nodes = NULL)
}
\arguments{
\item{network}{A flow network}

\item{nodes}{tibble of flow nodes, or a subset of these}
}
\value{
[tibble][tibble::tibble-package]
}
\description{
Get edges tibble from flow network
}
