% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{flow_network}
\alias{flow_network}
\title{Create a tidygraph from flows and remove spurious edges in the process.}
\usage{
flow_network(
  flows,
  spurious_thresholds = c(rate_o = 0.1, rate_d = 0.1),
  label_subgraphs = TRUE
)
}
\arguments{
\item{flows}{Asymptotic od flows}

\item{spurious_thresholds}{Remove edges whose attributes are lower than
the corresponding threshold.}
}
\description{
Create a tidygraph from flows and remove spurious edges in the process.
}
