% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{flow_nodes}
\alias{flow_nodes}
\title{Get nodes tibble from flow network}
\usage{
flow_nodes(network)
}
\arguments{
\item{network}{A flow network}
}
\value{
[tibble][tibble::tibble-package]
}
\description{
Get nodes tibble from flow network
}
