% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_corridor_set}
\alias{get_corridor_set}
\title{Compute corridor set from ordinary sequences}
\usage{
get_corridor_set(ord_sequences)
}
\arguments{
\item{ord_sequences}{tibble of ordinary sequences}
}
\value{
tibble of corridors
}
\description{
Compute corridor set from ordinary sequences
}
