% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_corridors}
\alias{get_corridors}
\title{Get corridors from observed sequences (wrapper)}
\usage{
get_corridors(observed_sequences, distances, ndays, G, min_rate, min_utility)
}
\arguments{
\item{observed_sequences}{observed trip sequences as strings}

\item{distances}{distances tibble}

\item{ndays}{total number of observed days}

\item{G}{ANPR flow network igraph}

\item{min_rate}{minimum observed daily rate}

\item{min_utility}{minimum route utility}
}
\value{
identified ANPR corridors tibble
}
\description{
Get corridors from observed sequences (wrapper)
}
