% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_datetime_labels}
\alias{get_datetime_labels}
\title{Get pretty time labels from breaks. Full format is applied to generate
the first two labels, whilst partial format is applied to all labels in
between.}
\usage{
get_datetime_labels(
  datetime_breaks,
  full_datetime_format = "\%y-\%m-\%d \%H:\%M",
  partial_datetime_format = "\%H:\%M"
)
}
\arguments{
\item{datetime_breaks}{input datetime breaks}

\item{full_datetime_format}{datetime format applied to first and last breaks}

\item{partial_datetime_format}{datetime format applied to elements in between}
}
\description{
Get pretty time labels from breaks. Full format is applied to generate
the first two labels, whilst partial format is applied to all labels in
between.
}
