% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_flows_od}
\alias{get_flows_od}
\title{Calculate flow rates for all OD pairs in the road network}
\usage{
get_flows_od(flows, flows_l, by_period = TRUE)
}
\arguments{
\item{flows}{A 'raw' flows dataset.}

\item{flows_l}{Flows at unique locations}

\item{by_period}{Whether to group data by time or not.}
}
\value{
flows_od [tibble][tibble::tibble-package]
}
\description{
Calculate flow rates for all OD pairs in the road network
}
