% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get the subset of the flow network which contains all nodes that are
neighbors to a list of nodes.}
\usage{
get_neighbors(network, nodes)
}
\arguments{
\item{network}{Flow network}

\item{nodes}{Character vector of node names}
}
\description{
Get the subset of the flow network which contains all nodes that are
neighbors to a list of nodes.
}
