% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_sink}
\alias{get_sink}
\title{Find and return the sink node of a graph}
\usage{
get_sink(g)
}
\arguments{
\item{g}{igraph}
}
\value{
sink node, if exists, null otherwise
}
\description{
Find and return the sink node of a graph
}
