% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_source}
\alias{get_source}
\title{Find and return the source node of a graph}
\usage{
get_source(g)
}
\arguments{
\item{g}{igraph}
}
\value{
source node, if exists, null otherwise
}
\description{
Find and return the source node of a graph
}
