% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_subgraphs}
\alias{get_subgraphs}
\title{Label nodes according to the subgraph, if any, that they belong in.}
\usage{
get_subgraphs(network)
}
\arguments{
\item{network}{A flow network}
}
\description{
Label nodes according to the subgraph, if any, that they belong in.
}
