% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_super_sequences}
\alias{get_super_sequences}
\title{Find the subset of ordinary sequences that are not a substring of any other sequence}
\usage{
get_super_sequences(ord_sequences)
}
\arguments{
\item{ord_sequences}{tibble of ordinary sequences}
}
\value{
tibble of sequences that are not a substring of any other sequence
}
\description{
Find the subset of ordinary sequences that are not a substring of any other sequence
}
