% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_labels}
\alias{get_time_labels}
\title{Get pretty time labels from breaks.}
\usage{
get_time_labels(time_breaks)
}
\arguments{
\item{time_breaks}{Trimmed flows tibble.}
}
\description{
Get pretty time labels from breaks.
}
