% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_total_flow}
\alias{get_total_flow}
\title{Get the total od flow in the network.}
\usage{
get_total_flow(flows_od, by_period = TRUE, ignore_sink_source = TRUE)
}
\arguments{
\item{flows_od}{Flows$od tibble.}

\item{by_period}{Whether to aggregate total flow by period.}

\item{ignore_sink_source}{Whether to include ods containing source and sink.}
}
\value{
flows_od [tibble][tibble::tibble-package]
}
\description{
Get the total od flow in the network.
}
