% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{is_simple_path}
\alias{is_simple_path}
\title{Does the simple path exist in G}
\usage{
is_simple_path(G, vs)
}
\arguments{
\item{G}{igraph}

\item{vs}{vertex sequence vector}
}
\value{
boolean
}
\description{
Does the simple path exist in G
}
