% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{is_valid_corridor}
\alias{is_valid_corridor}
\title{Determine if the input corridor is valid}
\usage{
is_valid_corridor(corridor, ord_seq = NULL)
}
\arguments{
\item{corridor}{tibble}
}
\value{
logical
}
\description{
Determine if the input corridor is valid
}
