% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{join_sequences}
\alias{join_sequences}
\title{Obtain the set of observed sequences from a trip dataset, the set of
theoretically possible sequences given an ANPR network and compute the
set intersection of the two.}
\usage{
join_sequences(G, trip_sequences, method = "e", sep = ",")
}
\arguments{
\item{G}{tidygraph of a ANPR network}

\item{trip_sequences}{tibble of observed trip sequences}

\item{method}{method to determine the set of possible trip sequences from
a ANPR network. If the network is relatively sparse and of small size then
this set can be compute exhaustively (method = "e"). Otherwise if the
network is too large and it becomes computationally infeasible to compute the
set, we use a heuristic (method = "h") to approximate the resulting subset.}

\item{sep}{character separating elements of a sequence encoded as a string}
}
\value{
a tibble of sequences and their counts
}
\description{
Obtain the set of observed sequences from a trip dataset, the set of
theoretically possible sequences given an ANPR network and compute the
set intersection of the two.
}
