% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{ordinary_sequences}
\alias{ordinary_sequences}
\title{Calculate the subset of trip sequences which are ordinary.}
\usage{
ordinary_sequences(
  sequences,
  min_rate = 30,
  min_utility = 0.7,
  .keep_cols = FALSE
)
}
\arguments{
\item{sequences}{a tibble of sequences (with columns 's', 'l', 'rate', 'utility')}

\item{min_rate}{minimum accepted daily observation rate: sequences observed
at a daily rate below this threshold are rejected.}

\item{min_utility}{minimum accepted route utility}

\item{.keep_cols}{whether to keep cols other than 's' and 'l'}
}
\value{
filtered tibble
}
\description{
Calculate the subset of trip sequences which are ordinary.
}
