% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_corridor_time}
\alias{plot_corridor_time}
\title{Plot a corridor's attribute as tiles over time. On the y axis,
a tile is generated for each segment of the corridor, for each time step,
where the bottom and top of the y axis represent the start and end of the
corridor, respectively. Segment levels are calculated from location levels.
Location levels should be adjusted before the function call.}
\usage{
plot_corridor_time(corridor_flows, fill_var, unite_sep = "-")
}
\arguments{
\item{corridor_flows}{corridor od flows [tibble][tibble::tibble-package]}

\item{fill_var}{var to fill geom_tiles with}

\item{unite_sep}{separator used to unite origin and
destination into single od character}
}
\value{
ggplot
}
\description{
Plot a corridor's attribute as tiles over time. On the y axis,
a tile is generated for each segment of the corridor, for each time step,
where the bottom and top of the y axis represent the start and end of the
corridor, respectively. Segment levels are calculated from location levels.
Location levels should be adjusted before the function call.
}
