% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_corridor_time_speed}
\alias{plot_corridor_time_speed}
\title{Plot a corridor's speed as tiles over time. On the y axis,
a tile is generated for each segment of the corridor, for each time step,
where the bottom and top of the y axis represent the start and end of the
corridor, respectively. Segment levels are calculated from location levels.
Location levels should be adjusted before the function call.}
\usage{
plot_corridor_time_speed(
  corridor_flows,
  fill_var = .data$mean_speed,
  unite_sep = "-",
  scale_fill_name = "Mean speed",
  scale_gradientn_colours = c("#E46972", "#EFBD7E", "#5B8E54"),
  scale_limits = c(0, 80)
)
}
\arguments{
\item{corridor_flows}{corridor od flows [tibble][tibble::tibble-package]}

\item{fill_var}{var to fill geom_tiles with}

\item{unite_sep}{separator used to unite origin and
destination into a single character}

\item{scale_fill_name}{name of fill scale}

\item{scale_gradientn_colours}{colours used to generate
the speed fill gradient}

\item{scale_limits}{fill scale limits}
}
\value{
ggplot
}
\description{
Plot a corridor's speed as tiles over time. On the y axis,
a tile is generated for each segment of the corridor, for each time step,
where the bottom and top of the y axis represent the start and end of the
corridor, respectively. Segment levels are calculated from location levels.
Location levels should be adjusted before the function call.
}
