% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_demand_l}
\alias{plot_demand_l}
\title{Plot traffic demand over time, grouped by location.}
\usage{
plot_demand_l(
  flows_l,
  time_breaks = NULL,
  point_alpha = 0.5,
  point_size = 0.75,
  title = NA
)
}
\arguments{
\item{flows_l}{Trimmed flows tibble.}

\item{point_alpha}{Alpha parameter of point layer}

\item{point_size}{Size parameter of point layer}

\item{flows}{Trimmed flows tibble.}
}
\description{
Plot traffic demand over time, grouped by location.
}
