% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_demand_od}
\alias{plot_demand_od}
\title{Plot traffic demand over time, grouped by od pair.}
\usage{
plot_demand_od(
  flows_od,
  time_breaks = NULL,
  title = NULL,
  point_alpha = 0.5,
  point_size = 0.75,
  line_size = 0.5
)
}
\arguments{
\item{flows_od}{Trimmed flows$od tibble.}
}
\description{
Plot traffic demand over time, grouped by od pair.
}
