% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot traffic demand over time, grouped by location.}
\usage{
plot_map(
  spatial,
  flows = NULL,
  add_primary = TRUE,
  add_arterial = FALSE,
  add_paths = TRUE,
  add_locations = TRUE,
  color_primary = munsell::mnsl(primary_col),
  color_arterial = munsell::mnsl(arterial_col),
  color_locations = munsell::mnsl(cameras_col),
  color_paths = munsell::mnsl(flows_col),
  size_primary = 0.2,
  size_arterial = 0.1,
  size_paths = 0.5,
  size_locations = 3,
  aes_color_locations = "",
  locations_palette = "Set2",
  aes_color_flows = ""
)
}
\arguments{
\item{spatial}{List of spatial features.}

\item{flows}{Optional summarised flows$l tibble.}

\item{add_primary}{Whether to plot primary network.}

\item{add_arterial}{Whether to plot arterial network.}

\item{add_paths}{Whether to plot shortest paths between locations.}

\item{color_primary}{Color of primary network.}

\item{color_arterial}{Color of arterial network.}

\item{color_locations}{Color of locations.}

\item{color_paths}{Color of paths.}

\item{size_primary}{Line size of the primary network.}

\item{size_arterial}{Line size of the arterial network.}

\item{size_paths}{Line size of pahts.}

\item{size_locations}{Size of locations.}

\item{aes_color_locations}{Color aesthetic to use for locations as
a character.}

\item{locations_palette}{Brewer palette for locations when used together
with aes_color_locations.}

\item{aes_color_flows}{Color aesthetic to use for paths as a character.}
}
\description{
Plot traffic demand over time, grouped by location.
}
