% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map_pairs}
\alias{plot_map_pairs}
\title{Plot a spatial map for every od pair in a given flow network.}
\usage{
plot_map_pairs(spatial, network, ...)
}
\arguments{
\item{spatial}{list of spatial features}

\item{network}{flow network}

\item{...}{parameters to be passed to plot_map}
}
\description{
Plot a spatial map for every od pair in a given flow network.
}
