% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_small_network}
\alias{plot_small_network}
\title{Plot flow network.}
\usage{
plot_small_network(
  network,
  include_source_sink = TRUE,
  graph_layout = "sugiyama",
  edge_label = NULL,
  node_label = .data$name,
  node_fill = .data$subgraph,
  edge_arrow = ggplot2::arrow(length = ggplot2::unit(3, "mm")),
  edge_start_cap = ggraph::circle(5, "mm"),
  edge_end_cap = ggraph::circle(5, "mm"),
  edge_angle_calc = "along",
  edge_label_dodge = ggplot2::unit(2.5, "mm"),
  edge_width = 1,
  edge_alpha = 0.75,
  num_accuracy = 0.1,
  num_scale = 1,
  label_size = 5,
  label_color = "black",
  source_sink_label_fill = "grey70"
)
}
\arguments{
\item{network}{Flow network.}

\item{include_source_sink}{Whether to include source and sink nodes.}

\item{graph_layout}{ggraph layout parameter}

\item{edge_label}{Optionally label edges according to this aesthetic}

\item{node_label}{Label nodes aesthetic}

\item{node_fill}{Fill nodes aesthetic}

\item{edge_arrow}{Edge arrow}

\item{edge_start_cap}{Arrow distance from starting node}

\item{edge_end_cap}{Arrow distance from ending node}

\item{edge_angle_calc}{Edges label angle parameter}

\item{edge_label_dodge}{Edges label dodge parameter}

\item{edge_width}{Edge width}

\item{edge_alpha}{Edge alpha}

\item{num_accuracy}{Number format accuracy of edge labels.}

\item{num_scale}{Multiplicative factor applied to edge labels.}

\item{label_size}{Node label size}

\item{label_color}{Font color of node labels}

\item{source_sink_label_fill}{Special fill color for source and sink nodes.}
}
\description{
Plot flow network.
}
