% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_speed_od}
\alias{plot_speed_od}
\title{Plot traffic speed over time, grouped by od pair.}
\usage{
plot_speed_od()
}
\arguments{
\item{flows}{Trimmed flows tibble.}
}
\description{
Plot traffic speed over time, grouped by od pair.
}
