% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_ts_matrix}
\alias{plot_ts_matrix}
\title{Matrix plot of traffic}
\usage{
plot_ts_matrix(
  flows_od,
  type = "demand",
  add_points = FALSE,
  aes_color = "",
  include_source_sink = TRUE,
  datetime_format = "\%d \%b \%Hh",
  ...
)
}
\arguments{
\item{flows_od}{flows_od tibble}

\item{type}{type}

\item{add_points}{whether to include point layer}

\item{aes_color}{color aesthetic}

\item{include_source_sink}{whether to include either source or sink nodes
in demand plot depending on out or incoming flow}

\item{datetime_format}{overriding datetime_format in individual plots}

\item{...}{further parameters passed on to plot_demand_od and plot_speed_od}
}
\description{
Matrix plot of traffic
}
