% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{reduce_sequences}
\alias{reduce_sequences}
\title{Sum the number of occurrences across each sequence.}
\usage{
reduce_sequences(seqs)
}
\arguments{
\item{seqs}{a tibble of sequences with columns "s","l","n"}
}
\value{
the same tibble with summarised counts
}
\description{
Sum the number of occurrences across each sequence.
}
