% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{scale_fill_speed}
\alias{scale_fill_speed}
\title{Sensible defaults for scale_fill_gradientn when using speed as fill aesthetic.}
\usage{
scale_fill_speed(
  p,
  name = "Mean speed",
  limits = c(0, 80),
  gradientn_colours = c("#E46972", "#EFBD7E", "#5B8E54")
)
}
\arguments{
\item{p}{a spacetime ggplot}

\item{name}{name of fill scale}

\item{limits}{fill scale limits}

\item{gradientn_colours}{colours used to generate
the speed fill gradient}
}
\value{
ggplot
}
\description{
Sensible defaults for scale_fill_gradientn when using speed as fill aesthetic.
}
