% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_first}
\alias{seq_first}
\title{Get first element of a string sequence}
\usage{
seq_first(s, ...)
}
\arguments{
\item{s}{string sequence}

\item{...}{arguments passed to seq_split}
}
\value{
character
}
\description{
Get first element of a string sequence
}
