% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{seq_last}
\alias{seq_last}
\title{Get last element of a string sequence}
\usage{
seq_last(s, ...)
}
\arguments{
\item{s}{string sequence}

\item{...}{arguments passed to seq_split}
}
\value{
character
}
\description{
Get last element of a string sequence
}
