% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{seq_split}
\alias{seq_split}
\title{Split string sequence by sep}
\usage{
seq_split(s, sep = ",")
}
\arguments{
\item{s}{string sequence}

\item{sep}{separator character}
}
\value{
character
}
\description{
Split string sequence by sep
}
