% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{small_network_plots}
\alias{small_network_plots}
\title{Plot flow network.}
\usage{
small_network_plots(network, ...)
}
\arguments{
\item{network}{Flow network.}

\item{...}{Parameters passed to plot_small_network}
}
\description{
Plot flow network.
}
