% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{spacetime_plotlist}
\alias{spacetime_plotlist}
\title{Facet a corridor od flows tibble, e.g. by date, and generate a space time
plot for each. This function was written as an alternative to the use of
facet_wrap which was behaving awkwardly with geom_tile.}
\usage{
spacetime_plotlist(corridor_flows, fill_var, facet_by, ...)
}
\arguments{
\item{corridor_flows}{corridor od flows [tibble][tibble::tibble-package]}

\item{fill_var}{var to fill geom_tiles with}

\item{facet_by}{var whose distinct values are used to facet the flows tibble}

\item{...}{further parameters passed to plot_spacetime}
}
\value{
list of ggplots
}
\description{
Facet a corridor od flows tibble, e.g. by date, and generate a space time
plot for each. This function was written as an alternative to the use of
facet_wrap which was behaving awkwardly with geom_tile.
}
