% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_multiple_time_steps}
\alias{stop_if_multiple_time_steps}
\title{Check whether time column exists and throw error}
\usage{
stop_if_multiple_time_steps(flows_od)
}
\arguments{
\item{flows_od}{OD flows tibble}
}
\description{
Check whether time column exists and throw error
}
