% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{subgraph_labels}
\alias{subgraph_labels}
\title{Label nodes according to the subgraph, if any, which they belong in.}
\usage{
subgraph_labels(network, ignore_source_sink = TRUE)
}
\arguments{
\item{network}{A flow network}

\item{ignore_source_sink}{Whether to ignore source and sink nodes in label
calculation.}
}
\value{
[tbl_graph]
}
\description{
Label nodes according to the subgraph, if any, which they belong in.
}
