% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{top_flows}
\alias{top_flows}
\title{Get the subset of flows whose OD pairs account for the top
p proportion of total or period-wise observed traffic flow.
Returns all od pairs up to the first od pair that exceeds p.}
\usage{
top_flows(flows_od, p = 0.9, by_period = TRUE, ignore_sink_source = TRUE)
}
\arguments{
\item{flows_od}{Flows$od tibble.}

\item{p}{Desired proportion of total traffic.}

\item{by_period}{Whether to aggregate total flow by period.}

\item{ignore_sink_source}{Whether to include ods containing source and sink.}
}
\description{
Get the subset of flows whose OD pairs account for the top
p proportion of total or period-wise observed traffic flow.
Returns all od pairs up to the first od pair that exceeds p.
}
