% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_st_crop}
\alias{try_st_crop}
\title{Check whether time column exists and throw error}
\usage{
try_st_crop(sf_tibble, bbox, sf_name)
}
\arguments{
\item{sf_tibble}{a simple features tibble}

\item{bbox}{a st_bbox object}

\item{sf_name}{the name of the tibble to be used if st_crop fails}
}
\description{
Check whether time column exists and throw error
}
