% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{utility_loss}
\alias{utility_loss}
\title{Calculate the utility loss of each sequence, given the shortest path
distance between each pair of locations.}
\usage{
utility_loss(sequences, distances, .keep_distances = FALSE)
}
\arguments{
\item{sequences}{a tibble of sequences (with columns 's', 'l')}

\item{distances}{a tibble of distances (with columns 'o', 'd', 'distance')}

\item{.keep_distances}{keep calculated distances as columns}
}
\value{
tibble with extra column 'uloss'
}
\description{
Calculate the utility loss of each sequence, given the shortest path
distance between each pair of locations.
}
