% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{vec_all_paths}
\alias{vec_all_paths}
\title{For each Search for all simple paths from source to sink. This is a wrapper around
igraph::all_simple_paths that tidys up the results into a single tibble.}
\usage{
vec_all_paths(Gs)
}
\arguments{
\item{Gs}{a list of graphs}
}
\value{
a tibble
}
\description{
For each Search for all simple paths from source to sink. This is a wrapper around
igraph::all_simple_paths that tidys up the results into a single tibble.
}
