% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_cols}
\alias{assert_cols}
\title{Simple wrapper around assertthat to assert expected column names}
\usage{
assert_cols(tib, required_cols)
}
\arguments{
\item{tib}{a tibble}

\item{required_cols}{character vector of columns expected to exist in tib}
}
\description{
Simple wrapper around assertthat to assert expected column names
}
