% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{crop_spatial}
\alias{crop_spatial}
\title{Crop spatial features based on od pairs.}
\usage{
crop_spatial(
  pairs,
  spatial,
  arterial_highway = "residential",
  bbox_margin = c(-500, -500, 500, 500)
)
}
\arguments{
\item{pairs}{OD pairs tibble.}

\item{spatial}{List of spatial features.}

\item{arterial_highway}{Only keep arterial edges with this value in the
highway attribute.}

\item{bbox_margin}{Spatial margin added to flows bounding box.}
}
\value{
[list]
}
\description{
Crop spatial features based on od pairs.
}
