% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dayperiod}
\alias{dayperiod}
\title{Categorise time into its corresponding period of the day}
\usage{
dayperiod(t, bounds = c(6, 10, 16, 21))
}
\arguments{
\item{t}{time vector}

\item{bounds}{daily time boundaries}
}
\value{
vector of corresponding periods# of the day (categorical)
}
\description{
Categorise time into its corresponding period of the day
}
