% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{flow_network}
\alias{flow_network}
\title{Create a tidygraph from flows and remove spurious edges in the process.}
\usage{
flow_network(
  flows,
  spurious_if_below = c(rate_o = 0.1, rate_d = 0.1),
  label_subgraphs = TRUE,
  subgraphs_ignore_source_sink = TRUE,
  names_as_factors = TRUE,
  node_levels = NULL
)
}
\arguments{
\item{flows}{Asymptotic od flows}

\item{spurious_if_below}{Remove edges whose attributes are lower than
the corresponding threshold.}

\item{label_subgraphs}{Whether to label subgraphs as a result of removing
spurious edges.}

\item{subgraphs_ignore_source_sink}{Whether to ignore source and sink
nodes in label calculation.}

\item{names_as_factors}{Whether to convert node names to factor.}

\item{node_levels}{Optionally set the levels of node factor.}
}
\value{
[tbl_graph]
}
\description{
Create a tidygraph from flows and remove spurious edges in the process.
}
