% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{flow_network}
\alias{flow_network}
\title{Create a tidygraph from flows and remove spurious edges in the process.}
\usage{
flow_network(
  flows,
  spurious_if_below = c(rate_o = 0.1, rate_d = 0.1),
  label_subgraphs = TRUE,
  keep_levels = TRUE
)
}
\arguments{
\item{flows}{Asymptotic od flows}

\item{spurious_if_below}{Remove edges whose attributes are lower than
the corresponding threshold.}

\item{label_subgraphs}{Whether to label subgraphs as a result of removing
spurious edges.}

\item{keep_levels}{Whether nodes should keep the same factor levels as flows.}
}
\value{
[tbl_graph]
}
\description{
Create a tidygraph from flows and remove spurious edges in the process.
}
