% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_flows_l}
\alias{get_flows_l}
\title{Calculate flows at all unique locations of the road network}
\usage{
get_flows_l(flows)
}
\arguments{
\item{flows}{A 'raw' flows dataset.}
}
\value{
flows_l [tibble][tibble::tibble-package]
}
\description{
Calculate flows at all unique locations of the road network
}
