% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_flows_od}
\alias{get_flows_od}
\title{Calculate flow rates for all OD pairs in the road network}
\usage{
get_flows_od(flows, flows_l)
}
\arguments{
\item{flows}{A 'raw' flows dataset.}

\item{flows_l}{Flows at unique locations}
}
\value{
flows_od [tibble][tibble::tibble-package]
}
\description{
Calculate flow rates for all OD pairs in the road network
}
