% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{ordinary_sequences}
\alias{ordinary_sequences}
\title{Calculate the subset of trip sequences which are ordinary.}
\usage{
ordinary_sequences(
  sequences,
  min_drate = 30,
  max_uloss = 0.25,
  .keep_cols = FALSE
)
}
\arguments{
\item{sequences}{a tibble of sequences (with columns 's', 'l', 'drate', 'uloss')}

\item{min_drate}{minimum accepted daily observation rate: sequences observed
at a daily rate below this threshold are rejected.}

\item{max_uloss}{maximum accepted utility loss: sequences with utility loss
above this threshold are rejected.}

\item{.keep_cols}{whether to keep cols other than 's' and 'l'}
}
\value{
filtered tibble
}
\description{
Calculate the subset of trip sequences which are ordinary.
}
