% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_demand_l}
\alias{plot_demand_l}
\title{Plot traffic demand over time, grouped by location.}
\usage{
plot_demand_l(
  flows_l,
  time_breaks = NULL,
  point_alpha = 0.5,
  point_size = 0.75,
  line_size = 0.5,
  out_flow = TRUE,
  include_source_sink = FALSE
)
}
\arguments{
\item{flows_l}{Trimmed flows$l tibble.}

\item{time_breaks}{Breaks for scale_x_datetime}

\item{point_alpha}{Alpha parameter of point layer}

\item{point_size}{Size parameter of point layer}

\item{line_size}{Size parameter of line layer}

\item{out_flow}{Whether to plot outgoing or incoming flow}

\item{include_source_sink}{Whether to include either source or sink nodes
in graph depending on out or incoming flow.}
}
\description{
Plot traffic demand over time, grouped by location.
}
