% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_demand_od}
\alias{plot_demand_od}
\title{Plot traffic demand over time}
\usage{
plot_demand_od(
  flows_od,
  time_breaks = NULL,
  add_points = TRUE,
  aes_color = "od",
  point_alpha = 0.5,
  point_size = 0.75,
  line_size = 0.5,
  include_source_sink = FALSE
)
}
\arguments{
\item{flows_od}{flows_od tibble.}

\item{time_breaks}{breaks for scale_x_datetime}

\item{add_points}{whether to include point layer}

\item{aes_color}{color aesthetic}

\item{point_alpha}{alpha parameter of point layer}

\item{point_size}{size parameter of point layer}

\item{line_size}{size parameter of line layer}

\item{include_source_sink}{include or exclude edges from source or to sink
nodes}
}
\description{
Plot traffic demand over time
}
