% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_demand_od}
\alias{plot_demand_od}
\title{Plot traffic demand over time, grouped by od pair.}
\usage{
plot_demand_od()
}
\arguments{
\item{flows}{Trimmed flows tibble.}
}
\description{
Plot traffic demand over time, grouped by od pair.
}
