% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_demand_od}
\alias{plot_demand_od}
\title{Plot traffic demand over time, grouped by od pair.}
\usage{
plot_demand_od(
  flows_od,
  time_breaks = NULL,
  point_alpha = 0.5,
  point_size = 0.75,
  line_size = 0.5,
  include_source_sink = FALSE
)
}
\arguments{
\item{flows_od}{Trimmed flows$od tibble.}

\item{time_breaks}{Breaks for scale_x_datetime}

\item{point_alpha}{Alpha parameter of point layer}

\item{point_size}{Size parameter of point layer}

\item{line_size}{Size parameter of line layer}

\item{include_source_sink}{Include or exclude edges from source or to sink
nodes.}
}
\description{
Plot traffic demand over time, grouped by od pair.
}
