% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot traffic demand over time, grouped by location.}
\usage{
plot_map(
  spatial,
  flows = NULL,
  add_primary = TRUE,
  add_arterial = FALSE,
  add_paths = TRUE,
  add_locations = TRUE,
  color_primary = munsell::mnsl(primary_col),
  color_arterial = munsell::mnsl(arterial_col),
  color_locations = munsell::mnsl(cameras_col),
  color_paths = munsell::mnsl(flows_col),
  size_primary = 0.2,
  size_arterial = 0.1,
  size_paths = 0.5,
  size_locations = 3,
  aes_color_locations = "",
  locations_palette = "Set2",
  aes_color_flows = ""
)
}
\arguments{
\item{spatial}{list of spatial features}

\item{flows}{a flows_od tibble}

\item{add_primary}{whether to plot primary network}

\item{add_arterial}{whether to plot arterial network}

\item{add_paths}{whether to plot shortest paths between locations}

\item{add_locations}{whether to plot observed locations}

\item{color_primary}{color of primary network}

\item{color_arterial}{color of arterial network}

\item{color_locations}{color of locations}

\item{color_paths}{color of paths}

\item{size_primary}{line size of the primary network}

\item{size_arterial}{line size of the arterial network}

\item{size_paths}{line size of pahts}

\item{size_locations}{size of locations}

\item{aes_color_locations}{color aesthetic to use for locations as
a character}

\item{locations_palette}{brewer palette for locations when used together
with aes_color_locations}

\item{aes_color_flows}{color aesthetic to use for paths as a character}
}
\description{
Plot traffic demand over time, grouped by location.
}
