% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_small_network}
\alias{plot_small_network}
\title{Plot flow network.}
\usage{
plot_small_network(
  flows_od,
  ignore_source_sink = FALSE,
  num_accuracy = 0.1,
  num_scale = 1,
  graph_layout = "auto",
  aes_edge_color = "",
  aes_edge_label = "flow",
  edge_angle_calc = "along",
  edge_label_dodge = ggplot2::unit(2.5, "mm"),
  edge_arrow = ggplot2::arrow(length = ggplot2::unit(1.5, "mm")),
  edge_start_cap = ggraph::circle(1, "mm"),
  edge_end_cap = ggraph::circle(2, "mm"),
  edge_width = 1,
  edge_alpha = 0.75,
  node_size = 4,
  node_color = munsell::mnsl("5BG 6/4"),
  aes_node_label = "name",
  node_label_nudge_y = 0.1,
  node_label_nudge_x = 0
)
}
\arguments{
\item{flows_od}{Summarised or windowed od flows tibble.}

\item{ignore_source_sink}{Whether to ignore source and sink nodes.}

\item{num_accuracy}{Number format accuracy of edge labels.}

\item{num_scale}{Multiplicative factor applied to edge labels.}

\item{graph_layout}{ggraph layout parameter}

\item{aes_edge_color}{Optionally color edges according to this aesthetic}

\item{aes_edge_label}{Optionally label edges according to this aesthetic}

\item{edge_angle_calc}{Edges label angle parameter}

\item{edge_label_dodge}{Edges label dodge parameter}

\item{edge_arrow}{Edge arrow}

\item{edge_start_cap}{Arrow distance from starting node}

\item{edge_end_cap}{Arrow distance from ending node}

\item{edge_width}{Edge width}

\item{edge_alpha}{Edge alpha}

\item{node_size}{Node size}

\item{node_color}{Node color}

\item{aes_node_label}{Optionally label nodes according to this aesthetic}

\item{node_label_nudge_y}{Shift amount applied node labels on the y axis}

\item{node_label_nudge_x}{Shift amount applied node labels on the x axis}
}
\description{
Plot flow network.
}
