% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_small_network}
\alias{plot_small_network}
\title{Plot flow network.}
\usage{
plot_small_network(
  flows_od,
  ignore_source_sink = FALSE,
  num_accuracy = 0.1,
  num_scale = 1,
  graph_layout = "auto",
  aes_edge_color = "",
  aes_edge_label = "flow",
  edge_angle_calc = "along",
  edge_label_dodge = ggplot2::unit(2.5, "mm"),
  edge_arrow = arrow(length = ggplot2::unit(1.5, "mm")),
  edge_start_cap = circle(1, "mm"),
  edge_end_cap = circle(2, "mm"),
  edge_width = 1,
  edge_alpha = 0.75,
  node_size = 4,
  node_color = munsell::mnsl("5BG 6/4"),
  aes_node_label = "name",
  node_label_nudge_y = 0.1,
  node_label_nudge_x = 0
)
}
\arguments{
\item{flows_od}{Summarised or windowed od flows tibble.}
}
\description{
Plot flow network.
}
