% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_speed_od}
\alias{plot_speed_od}
\title{Plot traffic speed over time, grouped by od pair.}
\usage{
plot_speed_od(
  flows_od,
  time_breaks = NULL,
  add_points = TRUE,
  aes_color = "od",
  point_alpha = 0.5,
  point_size = 0.75,
  add_ribbon = TRUE,
  ribbon_alpha = 0.4,
  ribbon_fill = "grey90",
  line_size = 0.5
)
}
\arguments{
\item{flows_od}{trimmed flows$od tibble}

\item{time_breaks}{breaks for scale_x_datetime}

\item{add_points}{whether to include point layer}

\item{aes_color}{color aesthetic}

\item{point_alpha}{alpha parameter of point layer}

\item{point_size}{size parameter of point layer}

\item{add_ribbon}{whether to add ribbon representing deviation from the mean}

\item{ribbon_alpha}{alpha parameter of ribbon layer}

\item{ribbon_fill}{fill parameter of ribbon layer}

\item{line_size}{size parameter of line layer}
}
\description{
Plot traffic speed over time, grouped by od pair.
}
