% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{subgraph_labels}
\alias{subgraph_labels}
\title{Label nodes according to the subgraph, if any, which they belong in.}
\usage{
subgraph_labels(network)
}
\arguments{
\item{network}{A flow network}
}
\value{
[tbl_graph]
}
\description{
Label nodes according to the subgraph, if any, which they belong in.
}
